ml/psychoscore/requirements.txt
# Core ML dependencies
torch>=2.1.0
transformers>=4.36.0
datasets>=2.15.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0

# MIDI processing
miditok>=3.0.0
mido>=1.3.0
pretty_midi>=0.2.10

# Audio processing (optional)
librosa>=0.10.0
soundfile>=0.12.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Inference server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Evaluation
scikit-learn>=1.3.0
matplotlib>=3.8.0

# Development
pytest>=7.4.0
black>=23.0.0
midiutil
